Favor Report a Problem to report errors
I don't think we use Connect for VC++ anymore. If that's the case, please remove it entirely. 

I've moved Report a Problem to the top to emphasize that it's the preferred option for bug reporting. I've also linked to the Report a Problem page (using the wrong style of docs.microsoft.com link).